feat(series): obtain info from yt playlist automatizated, cookies and chapters

+++ src/app/pages/series/series.component.ts
@@ -2,7 +2,7 @@ import { Component } from '@angular/core';
 import { NavbarComponent } from '../../components/navbar/navbar.component';
 import { YtVideoEmbedComponent } from '../../components/yt-video-embed/yt-video-embed.component';
 import { InputComponent } from '../../components/input/input.component';
-import { PlaylistDTO } from '../../models/serie.model';
+import { PlaylistDTO, StateEnum } from '../../models/serie.model';
 import { PlaylistService } from '../../services/playlist.service';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
@@ -14,6 +14,10 @@ import {
   MatDialogConfig,
   MatDialogModule,
 } from '@angular/material/dialog';
+import { ConfirmActionComponent } from '../../modals/confirm-action/confirm-action.component';
+import { Observable } from 'rxjs/internal/Observable';
+import { FilterSeriesComponent } from '../../modals/filter-series/filter-series.component';
+import { SpinnerComponent } from '../../components/spinner/spinner.component';
 
 @Component({
   selector: 'app-series',
@@ -27,6 +31,7 @@ import {
     CommonModule,
     FormsModule,
     CreateEditPlaylistComponent,
+    SpinnerComponent
   ],
   templateUrl: './series.component.html',
   styleUrl: './series.component.scss',
@@ -36,6 +41,7 @@ export class SeriesComponent {
   public playlists: PlaylistDTO[] = [];
   private subscription!: Subscription;
   private dialogConfig!: MatDialogConfig;
+  public isLoading: boolean = true;
 
   constructor(
     private playlistService: PlaylistService,
@@ -47,24 +53,38 @@ export class SeriesComponent {
     this.subscription = this.playlistService.playlists$.subscribe(
       (playlists) => {
         this.playlists = playlists;
+        this.isLoading = false;
+        this.isCreatingPath();
       },
       (error) => {
         console.error(error);
       }
     );
-    this.loadPlaylists();
 
     this.dialogConfig = new MatDialogConfig();
+    this.dialogConfig.width = '90%';
+    this.dialogConfig.maxWidth = '820px';
+    this.dialogConfig.height = 'auto';
+    this.dialogConfig.maxHeight = '80vh';
     this.dialogConfig.data = { playlist: null };
-    this.updateDialogConfig(false);
     this.dialogConfig.enterAnimationDuration = '500ms';
     this.dialogConfig.exitAnimationDuration = '300ms';
     this.dialogConfig.hasBackdrop = true;
     this.dialogConfig.autoFocus = true;
+
+    // this.openCreateDialog();
   }
 
   ngOnDestroy() {
     this.subscription.unsubscribe();
+  }
+
+  private isCreatingPath(): void {
+    const currentUrl = this.router.url;
+    if (currentUrl === '/series/nueva') {
+      this.router.navigate(['/series']);
+      this.openCreateDialog();
+    }
   }
 
   public loadPlaylists() {
@@ -72,12 +92,35 @@ export class SeriesComponent {
   }
 
   public openCreateDialog() {
-    this.updateDialogConfig(false);
+    const newOrderView = this.playlists.length + 1;
+    this.dialogConfig.data = { playlist: null, newOrderView: newOrderView };
 
     const dialogRef = this.dialog.open(
       CreateEditPlaylistComponent,
       this.dialogConfig
     );
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.playlistService.refreshPlaylists();
+      }
+    });
+  }
+
+  public openFilterDialog() {
+    this.dialogConfig.data = { playlist: null };
+
+    const dialogRef = this.dialog.open(FilterSeriesComponent, {
+      data: { playlist: null },
+      width: '90%',
+      maxWidth: '820px',
+      height: 'auto',
+      maxHeight: '80vh',
+      enterAnimationDuration: '500ms',
+      exitAnimationDuration: '300ms',
+      hasBackdrop: true,
+      autoFocus: true,
+    });
 
     dialogRef.afterClosed().subscribe((result) => {
       if (result) {
@@ -87,9 +130,8 @@ export class SeriesComponent {
   }
 
   public openEditDialog(playlist: PlaylistDTO) {
-    this.updateDialogConfig(true, playlist);
-
     this.dialogConfig.data = { playlist: playlist };
+
     const dialogRef = this.dialog.open(
       CreateEditPlaylistComponent,
       this.dialogConfig
@@ -102,16 +144,70 @@ export class SeriesComponent {
     });
   }
 
-  public deletePlaylist(identifier: string) {
-    this.playlistService.deletePlaylist(identifier).subscribe(
-      (response) => {
-        console.log(response);
+  public updatePlaylistState(playlist: PlaylistDTO) {
+    let newState: StateEnum;
+
+    switch (playlist.state) {
+      case StateEnum.WATCHED:
+        newState = StateEnum.WATCHING;
+        break;
+      case StateEnum.WATCHING:
+        newState = StateEnum.NOT_WATCHED;
+        break;
+      case StateEnum.NOT_WATCHED:
+        newState = StateEnum.WATCHED;
+        break;
+      default:
+        newState = StateEnum.WATCHED;
+        break;
+    }
+
+    const newStateName = newState;
+
+    playlist.state = newState;
+
+    this.playlistService.updatePlaylistState(playlist.shortName, newStateName).subscribe(
+      () => {
         this.playlistService.refreshPlaylists();
       },
-      (error) => {
-        console.error(error);
-      }
     );
+  }
+
+  public deletePlaylist(identifier: string) {
+    this.confirmDeleteDialog().subscribe((confirmation) => {
+      if (confirmation) {
+        this.playlistService.deletePlaylist(identifier).subscribe(
+          (response) => {
+            console.log(response);
+            this.playlistService.refreshPlaylists();
+          },
+          (error) => {
+            console.error(error);
+          }
+        );
+      }
+    });
+  }
+
+  private confirmDeleteDialog(): Observable<boolean> {
+    const dialogRef = this.dialog.open(ConfirmActionComponent, {
+      data: {
+        title: 'Confirmar eliminación',
+        text: '¿Estás seguro de que deseas eliminar esta playlist?',
+        confirmIcon: 'fa-solid fa-trash',
+        confirmText: 'Eliminar',
+      },
+      height: 'auto',
+      width: '90vw',
+      maxHeight: '250px',
+      maxWidth: '500px',
+      enterAnimationDuration: '500ms',
+      exitAnimationDuration: '300ms',
+      hasBackdrop: true,
+      autoFocus: true,
+    });
+
+    return dialogRef.afterClosed();
   }
 
   public navigate(link: string, openInNewTab: boolean): void {
@@ -130,26 +226,5 @@ export class SeriesComponent {
     return thumbnail == 'no-have-thumbmail'
       ? 'assets/imgs/no-have-thumbmail.png'
       : `data:image/jpeg;base64,${thumbnail}`;
-  }
-
-  private updateDialogConfig(isEdit: boolean = false, playlist?: PlaylistDTO) {
-    // Data
-    if (isEdit) {
-      this.dialogConfig.data = { playlist: PlaylistDTO };
-    } else {
-      this.dialogConfig.data = { playlist: null };
-    }
-    // Width and Height
-    if (window.innerWidth >= 1280) {
-      this.dialogConfig.width = '42.5%';
-      this.dialogConfig.height = '80%';
-    }
-    if (window.innerWidth < 1280) {
-      this.dialogConfig.width = '60%';
-    }
-    if (window.innerWidth < 768) {
-      this.dialogConfig.width = '90%';
-      this.dialogConfig.height = '90%';
-    }
   }
 }
